Added the bones of dual-player presenting as single player for left/right audio, needs debug

+++ RubyonRails/app/controllers/audio_files_controller.rb
@@ -11,82 +11,89 @@ class AudioFilesController < ApplicationController
     render 'debug/audio_testing'
   end
   # for playing the file
-  def play
-    # currently hardcoded for easy testing
-    audio_file = Rails.root.join('app', 'assets', 'audio', '1-pair Dichotic Digits, List 1_Left_HRTF.wav')
-    #TODO Make sure that the updated temp file is being player here (currently downloaded?) Need to un-hardcode this for further testing
-    #debug to check for file being uploaded etc
-    if File.exist?(audio_file)
-      send_file audio_file, type: 'audio/wav', disposition: 'inline'
-    else
-      Rails.logger.error "Audio file not found: #{audio_file}"
-      head :not_found
-    end
+
+  # For playing in the 'left ear' player
+  def play_left
+    play_audio('left')
+  end
+
+  # For playing in the 'right ear' player
+  def play_right
+    play_audio('right')
+  end
+
+  private
+
+  def play_audio(channel)
+    # finds the latest temp file to play
+
+    file = find_latest_temp_file(channel) || default_file(channel)
+    send_file file, type: 'audio/wav', disposition: 'inline'
+    
     # this is currently coded for .wav!!
   end
 
+  # helper for finding the most recent file change for the respective ear channel/player
+  def find_latest_temp_file(prefix)
+    Dir.glob(Rails.root.join('tmp', "#{prefix}_*.wav")).max_by { |f| File.mtime(f) }
+  end
+
+  # This will be controlling the default file for each page, will need futher work and specifics on each view. Right now it is hardcoded.
+  def default_file(channel)
+    #
+    Rails.root.join('app', 'assets', 'audio', "1-pair Dichotic Digits, List 1_#{channel}_HRTF.wav")
+  end
+  
   # for adjusting decibels
   # currently hard coded audio!!!
   def adjust
-    input_file = Rails.root.join('app', 'assets', 'audio', '1-pair Dichotic Digits, List 1_Left_HRTF.wav')
-    Rails.logger.debug "Input file path: #{input_file}"
-    Rails.logger.debug "Input file exists: #{File.exist?(input_file)}"
+
+    left_input = params[:channel] == 'left' ? find_latest_temp_file('left') : default_file('left')
+    right_input = params[:channel] == 'right' ? find_latest_temp_file('right') : default_file('right')
 
     # Generates a unique filename for the new file
-    output_filename = "adjusted_audio_#{Time.now.to_i}_#{rand(1000)}.wav"
+    # output_filename = "adjusted_audio_#{Time.now.to_i}_#{rand(1000)}.wav"
     # Create a temporary file for the output
-    output_file = Rails.root.join('tmp', output_filename)
+    #output_file = Rails.root.join('tmp', output_filename)
 
     #adjusting decibels by param, must be specified when the controller is called to change volume
     # ie POST /adjust_audo?decibel_change=10 would increase decibels by 10
     decibel_change = params[:decibel_change].to_f
+    channel = params[:channel]
 
-    # defining succes for debug
-    command = "ffmpeg -i \"#{input_file}\" -filter:a \"volume=#{decibel_change}dB\" \"#{output_file}\""
-    output = `#{command} 2>&1`
-    success = $?.success?
-    Rails.logger.debug "FFmpeg output: #{output}"
-    Rails.logger.debug "FFmpeg command success: #{success}"
+    left_output = process_audio(left_input, decibel_change, channel == 'left')
+    right_output = process_audio(right_input, decibel_change, channel == 'right')
 
-    #this is the actual ffmpeg command
-    #can be positive for increase, can pass negative for decrease
-    #TODO: Add ear distinction? Could be added views side potentially?
+      #this hooks into our method for debugging/checking decibel change
+      #original_analysis = analyze_audio(input_file)
+      left_analysis = analyze_audio(left_output)
+      right_analysis = analyze_audio(right_output)
 
-    # serves the adjusted audio file to the user's view page
-    # currently downloads the value
-    
-    #TODO Find a good place to clean up the temporary file later when done
-
-    #file change success debug
-    if success
+      Rails.logger.debug "Original audio analysis: #{original_analysis}"
+      Rails.logger.debug "Left adjusted audio analysis: #{left_analysis}"
+      Rails.logger.debug "Right adjusted audio analysis: #{right_analysis}"
 
       cleanup_old_temp_files
 
-      #this hooks into our method for debugging/checking decibel change
-      original_analysis = analyze_audio(input_file)
-      adjusted_analysis = analyze_audio(output_file)
-
-      Rails.logger.debug "Original audio analysis: #{original_analysis}"
-      Rails.logger.debug "Adjusted audio analysis: #{adjusted_analysis}"
-
-      # this is the actual portion where the file is sent to be output
-      send_file output_file, type: 'audio/wav', disposition: 'inline'
-    else
-      Rails.logger.error "FFmpeg command failed: #{command}"
-      head :internal_server_error
-    end
-  rescue => e
-    Rails.logger.error "Error in adjust method: #{e.message}"
-    Rails.logger.error e.backtrace.join("\n")
-    head :internal_server_error
-  ensure
-    # Makes sure to delete output file if there's an error
-    
-    File.delete(output_file) if File.exist?(output_file)
-    
+      render json: { 
+      left_url: url_for(controller: 'audio_files', action: 'play_left'),
+      right_url: url_for(controller: 'audio_files', action: 'play_right'),
+      left_analysis: left_analysis,
+      right_analysis: right_analysis
+      }
+      # TODO make this show in console
+      
   end
 
-
+  # new method for audio adjustment to free it from being part of adjust, easier to edit
+  def process_audio(input, decibel_change, apply_change)
+    # generates a random new file name
+    output = Rails.root.join('tmp', "#{File.basename(input, '.*')}_#{Time.now.to_i}_#{rand(1000)}.wav")
+    # changes the decibels based on the passed value for that new file
+    filter = apply_change ? "volume=#{decibel_change}dB" : 'anull'
+    system("ffmpeg", "-i", input.to_s, "-filter:a", filter, output.to_s)
+    output
+  end
     # This should hopefully allow for an AJAX request
     # this is a javascript responsea
     # this will likely bne needed for live update using buttons in the future
@@ -96,25 +103,37 @@ class AudioFilesController < ApplicationController
   
     # This is for testing that the decibels are being change appropriately!
     def analyze_audio(file_path)
+
       command = "ffmpeg -i \"#{file_path}\" -af volumedetect -f null /dev/null 2>&1"
+  
+  
+      Rails.logger.debug "Analysis command: #{command}"
+
       output = `#{command}`
+
       mean_volume = output.match(/mean_volume: ([-\d.]+) dB/)&.captures&.first
       max_volume = output.match(/max_volume: ([-\d.]+) dB/)&.captures&.first
+
+      # TODO: make sure this ruby hash is showing in console
       { mean_volume: mean_volume, max_volume: max_volume }
+
+
     end
 
     # Handles cleanup of temp files the site no longer needs to avoid file bloat
     def cleanup_old_temp_files
       temp_dir = Rails.root.join('tmp')
       threshold_time = 1.hour.ago
-    
-      Dir.glob(temp_dir.join('adjusted_audio_*')).each do |file|
-        if File.mtime(file) < threshold_time
-          File.delete(file)
-          Rails.logger.info "Deleted old temp file: #{file}"
-        end
+  
+      Dir.glob(temp_dir.join('left_*')).each do |file|
+        File.delete(file) if File.mtime(file) < threshold_time
+      end
+
+      Dir.glob(temp_dir.join('right_*')).each do |file|
+        File.delete(file) if File.mtime(file) < threshold_time
       end
     end
+  end
 end
 
 =begin
